pyproject.toml
[tool.poetry]
name = "communications-management-system"
version = "0.1.0"
description = "communications-management-system"
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
pre-commit = "^3.1.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.1"
django-filter = "^24.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
